move cache file
